Polish PR3 SPEED mode for review readiness.
Keep m/s computation in speed.py, dedupe model and min-speed constants,
reuse minimap transforms for gap-fill, drop unused homography helper,
and document SPEED in the soccer README.
